requirements.txt
psycopg2==2.9.11
pyrogram==2.0.106
tgcrypto
SQLAlchemy==2.0.32
python-dotenv==1.0.1
pytz==2024.1
APScheduler==3.10.4
google-generativeai==0.7.2
aiohttp==3.9.1
beautifulsoup4==4.12.2